requirements/base.txt
aiohttp==3.6.2
dataclasses==0.5; python_version<"3.7"
pyyaml==5.3
async_generator==1.10; python_version<"3.7"
python-jose==3.1.0
python-dateutil==2.8.1
yarl==1.4.2
aiodocker>=0.16.0
click==7.1.1
colorama==0.4.3; platform_system=="Windows"
humanize==2.0.0
# upgrade setup.py after pinning certtifi to new version
certifi==2019.11.28
psutil==5.7.0
typing_extensions==3.7.4.1
wcwidth==0.1.8
toml==0.10.0
requests==2.23.0
-e .
